Added circular buffering to VC Buffers
Added the input ports VC_circEN (circular mode enable), VC_circSel
(buffer select for circular mode), and VC_directEnq (direct enqueue to
VC - circular mode must be enabled for this to work) to the
virtual_channel.vhd.

Updated the TLM BlazeRouter.vhd to reflect changes in VC. Added the
signals rnaVcNorthCircEn, rnaVcNorthCircSel, and rnaVcNorthDirectEnq so
that the RNA can communicate with the updated VC (there is also a set
for east, south and west). RNA needs to be updated to communicate with
these ports