This commit has the following changes/additions:
+ Added new VCC related signals within the Control Unit to manage and
keep track of the number of packets being enqueued and dequeued from
each cell within the West VC. Eventually, as we test this out and
verify operation, we'll multiple the functionality to the three other
ports.

+ Framework is set within the new process to handle a circular shift of
packets. Have not fully implemented the logic to do so, but it's
anticipated that the next commit will have this portion working. Timing
will also be analyzed to see how long it takes to perform. Hopefully
it's negligible.

Signed Off: KVH