This commit has the following:
x_sch_departed_ack, where x = N,S,E,W was not being set to '1' once the
packet departed from the router to the ejection buffer after the
scheduling completed. This is an indicator back to the scheduler that
you may proceed to the next item in your list. Therefore, multiple
empty packets were being inserted into the ejection buffer until the
scheduler expired.

However, I'm still noticing the Control Packet entering the ejection
buffer to the PE. I guess there really is no benefit to the PE to see
this, because it's truly a property of routing, but until I dig deeper
under the hood of the VC and Switch code, I won't know what's causing
it. Hopefully, i'll have this corrected by the next commit. :)