This commit has the following:
Added the ability to log the test bench results to a local file and the
simulator console, simultaneously. "rt1_results_file.txt" is an example
of what each router is expected to spit out when injecting/ejecting
data. I'll probably add some more detail (at least when the packet is
ejected at the destination). The timing component should give me a
better idea of how to look at the packet traversing the NoC (albeit,
small NoC) and make assessments on my results.