This commit has the following:
Forgot to add in the sorter code for E,S, and W the check for all 1's,
which indicates a completed transfer. Those ports were scheduling those
jobs, but they're not valid. This has been corrected.

Playing with the test bench and the timing. Looking good so far. Need
to refine the test bench a little more to handle various scenarios, but
it looks like it's coming together. You can run the simulator and
observe the .txt file results.